Correcting particle support for paranoid compilation

+++ src/BallTree.h
@@ -41,7 +41,7 @@ public:
 //        printf("Building BallTree (%d) ...\n", n);
         nr.resize(n);
         if (n > 0) {
-            for (int i=0; i<n; ++i) nr[i] = i;
+            for (size_t i=0; i<n; ++i) nr[i] = i;
             build(0,n,-1);
         }
 //        printf("Done (%d)\n", tree.size());
